+++ src/main/java/site/doto/global/code/ErrorCode.java
@@ -1,11 +1,11 @@
-package site.doto.global.common.code;
+package site.doto.global.code;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum SuccessCode {
+public enum ErrorCode {
     ;
 
     private final int httpStatusCode;
