+++ src/main/java/site/doto/global/dto/ResponseDto.java
@@ -1,10 +1,10 @@
-package site.doto.global.common.dto;
+package site.doto.global.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
-import site.doto.global.common.code.ErrorCode;
-import site.doto.global.common.code.SuccessCode;
+import site.doto.global.code.ErrorCode;
+import site.doto.global.code.SuccessCode;
 
 @Data
 @AllArgsConstructor
