+++ src/main/java/site/doto/global/exception/CustomException.java
@@ -2,7 +2,7 @@ package site.doto.global.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import site.doto.global.common.code.ErrorCode;
+import site.doto.global.code.ErrorCode;
 
 @AllArgsConstructor
 @Getter
